Upgrades SQLAdmin version
This version uses google-api-client 1.23.0 instead of 1.22.0.
1.23.0 is already in use by other dependencies (e.g., everything at
google-cloud-java).
The fact that this dependency is pulling 1.22.0 is causing diamond
dependency issues in Spring Cloud GCP.
We had to manually set the version to be used on our side but,
ideally, this project should set the right version.
Better yet, this version should be obtained from dependencyManagement,
so we could use BOMs and not enforce any versions.