Update README.md
Added firewall configuration notes. It appears that the socket factory attempts to connect to port 3307 on the Cloud SQL instance, but that is not documented anywhere and is not obvious at all. The port number is currently hard-coded as a constant: `com.google.cloud.sql.core.SslSocketFactory.DEFAULT_SERVER_PROXY_PORT` and it took me a while to find it. I hope that the proposed documentation will make troubleshooting easier for other people.