chore: skip assembly for parent (#691)
Previously, attempts to run the assembly plugin would fail because the
parent was empty. Now that we've removed all dependencies from the
parent and placed them where they belong in the respective children
artifacts, the parent is empty. This commit updates the configuration of
the assembly step to skip the parent and only run for the children.