chore: prefer synchronous interfaces
This commit makes the SqlAdminApiFetcher use a synchronous interface. In
addition, the RSA key pair generation is done once at initialization
time of the core socket factory.

As part of the change, we create a ScheduledExecutorService at the core
socket factory level and decorate it with a listenable interface only
where needed. Otherewise, the ScheduledExecutorService is the primary
executor type.