fix: Fix refresh futures to avoid a hanging future when an api request fails during refresh. (#1319)
There was a subtle bug in the way we were using futures during the process to refresh the tokens and ephemeral
certificates. Occasionally an API request would fail, causing the refresh process to hang.

Now the code uses Futures.whenAllComplete() instead of Futures.whenAllSucceded() so that an api request error
will be propagated rather than ignored.