chore: add AccessTokenSupplier (#1329)
Between version v1.13 and v1.14 of google-auth-library-oauth2, a bug was fixed that changed the behavior of how downscope tokens are refreshed. Prior to the fix, the code could downscope an oauth2 token, then the refresh the downscoped token. After the fix, the code must first refresh the original oauth2 token, then downscope it again.

In order to incorporate this new behavior into our code, we needed to encapsulate the operation of refreshing and
downscoping a token. This change became too big for one PR, so I split the implementation into 3 PRs.

This PR introduces a new class and interface AccessTokenSupplier and DefaultAccessTokenSupplier along with extensive unit tests. This new class encapsulates the logic to refresh a token, duplicating the logic in CloudSqlInstance

PR #1330 refactors CloudSqlInstance and SqlAdminApiFetcher and tests to use the new DefaultAccessTokenSupplier, removing the duplicate refresh token code.

PR #1331 upgrades google-auth-library-oauth2 to the latest version.

Related to issue: #1294