refactor: Move refresh logic into Refresher for easier sharing with AlloyDb (#1615)
Create a new class Refresher which knows how to refresh a data object with an expiration date.
This moves the refresh algorithm out of CloudSqlInstance so that it is easier to share the refresh algorithm
implementation between this project and the AlloyDB Java Connector.

The api contract of CloudSqlInstance used by CoreSocketFactory does not change. CloudSqlInstance now has a private Refresher which it configures to keep InstanceData valid.

refresher =
    new Refresher<>(
        connectionName, // The name of the connector, in refresh log messages
        executor, // The executor
        // Supplier<ListenableFuture<InstanceData>> which retrieves fresh data.
        () ->  instanceDataSupplier.getInstanceData(
                      this.instanceName, this.accessTokenSupplier, this.authType, executor, keyPair),
        new RateLimitCalculator(minRefreshDelayMs)); // The rate limiter.

Another benefit: by encapsulating the refresh algorithm encapsulated in a separate class, it will be much easier to keep the AlloyDB Java Connector and the Cloud SQL Java Connector refresh implementations in sync. Without this refactor, we will need to copy the logic from CloudSqlInstance and paste it into DefaultConnectionInfoCache. Then, we will need to make A LOT of subtle changes. It will be very hard to keep these implementations in sync over time.