feat: Use standard TLS hostname validation for instances with DNS names. (#2125)
When the Cloud SQL Instance reports that it has a DNS Name, the connector will use standard TLS hostname 
validation when checking the server certificate. Now, the server's TLS certificate must contain a SAN record 
with the instance's DNS name.

The ConnectSettings API added a field dns_names which contains all of the valid DNS names for
an instance.

See also GoogleCloudPlatform/cloud-sql-go-connector#954