feat: Metadata Exchange support for the Java connector. (#2200)
This is the client-side implementation of the metadata exchange protocol for Cloud SQL instances. This will
allow connectors to pass metadata about the connection to the Cloud SQL instance while the connection
is being established.

In particular, this is a part of resolving the issues with the Auth Proxy connecting to MySQL instances that
use the caching_sha2_password extension. (See GoogleCloudPlatform/cloud-sql-proxy#2317)

This does not update the default MDX behavior yet. That will be added in #2207